firmware/*: break EFM32GG into a MAC and PHY
This defines a basic PHY interface so that alternate PHYs may eventually
be used with the EFM32GG MAC. The KSZ8091 starts to implement this PHY
interface. The EFM32GG MACPHY also no longer hardcodes the address of
the PHY. It is dynamically discovered at runtime, further allowing
alternate PHYs to be used in the future.