WIP: firmware/*: use RTFM
I'm having a bit of trouble getting the lifetimes to work with the types
from smoltcp. RTFM doesn't appear to allow the use of lifetimes on
resources, which leaves me to use 'static. This is a problem because the
underlying buffers cannot be defined 'static from within RTFM's init
routine.