tool: run openocd from within gdb
This is so much easier than having to run openocd in the background.
This also cleans up the OpenOCD config by making it a little more clear
that off/false aren't related to uart, and removing the unneded trace
frequency. The ITM output file is also being changed from a FIFO to a
regular file. This is because OpenOCD hangs until the other end of the
pipe is opened by itmdump, and it's a bit confusing when it happens
within gdb.