main: migrate to RTIC 1.0
The main reason for migrating is the task cancel/reschedule
functionality, which will allow network handling to be scheduled
correctly. Without that mechanism, it wasn't possible for a network
event to reschedule the handling before any existing scheduled
handling. With the new structure of 1.0 though, the logging was
allowed to be greatly simplified and now allows logging to happen from
within the init function! This is because the logger has been moved
into the mutable static (now a _real_ mutable static) which allows the
logger to remain in the same location of memory for it's lifetime.

I took the opportunity to restructure the socket handling
slightly. The bulk of the implementation has been moved into the
network module to help de-clutter main. The network scheduling has
also been scheduled for immediately after initialization
finishes. This will allow any initial network events to be
processed (there aren't currently at the moment, but they will soon
be).