*: use phy interrupts
Instead of immediately spawning the network handler, wait until the
link is up. This allows DHCP to complete more quickly because instead
of incurring the 10-second back-off after transmitting into the void,
the firmware waits until the PHY is ready to transmit.

There is a one-second delay between the link-up interrupt and the
network handler running. I'm not sure why this is needed, but I'll
have to dig into it later.