flyback: add initial board layout
I think I'm happy with this as-is for the first revision. For the
future though, I think it might be cleaner to move T2 and C16 to the
other side of T1. There were also a number of design rule exclusions
related to the clearance around VSSA and VDDA. I was shooting for a
spacing of 0.025" as advised by the reference manual, but a few of the
footprints were too small to accommodate.