efm32gg: remove LED configuration
This was always a bit of a hack to help with debugging. Neither the
MAC nor PHY should know anything about the LEDs. I tried making this
more generic with hooks, but it ended up a little unbalanced - the
LEDs were passed into the interrupt handler, but were stolen by the
hooks. In the long term, it might make more sense to move all of this
functionality to RX/TX start/stop hooks.