*: use network LED to show state
This has made it quite a bit easier to troubleshoot. The network LED
remains solid on when the network is uninitialized (this would only be
seen if the firmware panicked early on) and solid off when the network
is fully configured (currently, successful DHCP). The rest of the
time, the network LED uses an occulting flash pattern to indicate the
state:

  1. No link
  2. No DHCP

The priority level of the two flashing tasks was increased to avoid a
deadlock (and theoretically improve the UX). The following panic was
seen when a DHCP lease was obtained:

    panicked at 'spawning occult_network_led: ()', src/bin/passthru.rs:153:41