*: add support for RTT
This setup allows multiple loggers (ITM and RTT currently) to operate
in parallel. Eventually, I'd like to introduce dynamically-enabled
loggers and the ability to change the logging level at runtime.

The OpenOCD configuration assumes that the RTT control block will be
within the first 4 KiB of memory, which is not a great
assumption. Ideally, the linker script would place the control block
into a known location in memory, but that's going to require
modification to both cortex-m and rtt-target.