Reserve space for programs to be loaded
This 0.5 KiB is set aside for programs to be written and run from
memory. The long term vision is to use this space to run a sort of
bootloader which can then transfer larger programs to flash. I'm
hoping that this will speed up development, since the lower part of
the firmware isn't changing as much and so doesn't need to be
re-flashed constantly.