firmware/build: rework the build-and-run process
When OpenOCD is started with a configuration that also includes
commands to start RTT, the logging only works if the existing firmware
has configured RTT properly. The previous setup would run that
configuration and _then_ flash the program. This resulted in very
confusing behavior where the RTT output from the latest firmware might
not show up until OpenOCD was restarted - or more accurately, the RTT
setup commands were run again.

I decided to remove all of the commands from the OpenOCD config, which
makes a bit more sense in retrospect, and include them instead in the
run script. This new script is also set up to start streaming the RTT
logs immediately after flashing the firmware. This requires running
OpenOCD twice, back to back, but it seems to be working well enough.